cb - searchform component created
